Reduce home page buttons and menu items
Out of three buttons on the home page, this keeps only Register. About
and News are already present in the top menu, and it seems fine to
keep them only there to make the home page look cleaner.

Register, on the contrary, is an external link, so it seems more
fitting as a button rather than menu item, and consequently is removed
from the menu (even though the theme took care of marking it as
external).