Add basic schedule rendering
This adds a shortcode which uses hugo's ability to work with xml files
to render the xml schedule as a set of headings and paragraphs.

The schedule file remains stored in the static directory (so that it
will always be served), but it is additionally symlinked into
assets/xml/schedule.xml where it can be picked up as a "global
resource".

Use `{{< schedule >}}` in markdown files to insert the rendered
schedule.

https://gohugo.io/functions/transform/unmarshal/?search-input=time#working-with-xml