Add generated schedule to the schedule page
This restructures the schedule page a little, and replaces a list of
accepted talks with an autogenerated schedule (list of abstracts, with
times).

The schedule is included in the page via a custom shortcode (1), which
uses hugo's ability to parse xml data. It renders a schedule based on
the content of the xml file (which we also use for giggity).