Display abstracts within html details tag
To further address the concerns that the current schedule page is
unwieldy (#38), the schedule shortcode is updated so that the abstract
is shown inside html `<details>` tag.

Observation on the side: the XML file can use either "abstract" or
"description". Giggity will use either, but our hugo shortcode
currently only takes "abstract". This can be used as a mechanism to
avoid having abstracts for coffee breaks.