Improve XML consistency: coffee has description, not abstract
The XML file was inconsistent in using description vs abstract for
coffee (and lunch) breaks. The former sounds more appropriate, and the
file is adjusted accordingly.

Although Giggity can use either of those fields, it has a side-effect
for our hugo website, as we currently only take abstract and display
it in the schedule page. This became important as abstracts got folded
into html details (with summary saying "Abstract") in e1639cf.

Reported-by: Adina Wagner <adina.wagner@t-online.de>