Reorganize day2 schedule to fit Dataverse mini-symposium
The idea that to fit contributions of three people

- Phil Durbin (Dataverse intro)
- Jan Range (pydataverse)
- Oliver Bertuch (dataverse connection to distributed data stores)

into a 40min slot, including questions.

It is placed in front of the (shortened) unconference slot, to have the
chance to co-opt that for any discussions that may come from this
presentation, without having to budget dedicated time for it.

An abstract is yet to come.